+++ tests/models/opt/test_modeling_tf_opt.py
@@ -87,7 +87,7 @@ class TFOPTModelTester:
 
     def prepare_config_and_inputs_for_common(self):
         input_ids = ids_tensor([self.batch_size, self.seq_length - 1], self.vocab_size)
-        eos_tensor = tf.expand_dims(tf.constant([self.eos_token_id] * self.batch_size), 1)
+        eos_tensor = tf.expand_dims(tf.constant([self.eos_token_id] * self.batch_size, dtype=tf.int64), 1)
         input_ids = tf.concat([input_ids, eos_tensor], axis=1)
 
         config = self.config_cls(
